move helper functions to utils.ts
