+++ packages/core/src/change-matrix.ts
@@ -8,8 +8,7 @@ import {
   TooltipPosition,
   SvgIcons
 } from './interfaces';
-import { getAnchor } from './utils';
-import { getGeneralChartInteractions, getModeBarMessages } from './onboarding';
+import { getAnchor, getGeneralChartInteractions, getModeBarMessages } from './utils';
 
 
 export interface IOnboardingChangeMatrixSpec extends IOnboardingSpec {
