+++ packages/core/src/heatmap.ts
@@ -7,8 +7,7 @@ import {
   IOnboardingStage,
   SvgIcons
 } from './interfaces';
-import { getAnchor } from './utils';
-import { getGeneralChartInteractions, getModeBarMessages } from './onboarding';
+import { getAnchor, getGeneralChartInteractions, getModeBarMessages } from './utils';
 
 export interface IOnboardingHeatmapSpec extends IOnboardingSpec {
   chartTitle?: ISpecProp;
