+++ packages/core/src/horizon-graph.ts
@@ -7,8 +7,7 @@ import {
   defaultOnboardingStages,
   SvgIcons
 } from './interfaces';
-import { getAnchor } from './utils';
-import { getGeneralChartInteractions, getModeBarMessages } from './onboarding';
+import { getAnchor, getGeneralChartInteractions, getModeBarMessages } from './utils';
 
 export interface IOnboardingHorizonGraphSpec extends IOnboardingSpec {
   chartTitle?: ISpecProp;
