+++ packages/core/src/onboarding.ts
@@ -9,13 +9,11 @@ import {
   IOnboardingMessage,
   IOnboardingStage,
   NavigationAlignment,
-  SvgIcons
 } from './interfaces'
 import { v4 as uuidv4 } from 'uuid'
 import { get } from 'svelte/store'
 import { getMarkerInformation } from './components/getMarkerInformation'
 import { VisahoiState } from './components/state'
-import { getMarkerDomId } from './utils'
 
 let onboardingUI: OnboardingUI
 
@@ -351,32 +349,4 @@ export const setEditMode = (contextKey: string, value: boolean) => {
     console.error('No store for contextKey ', contextKey)
     return null
   }
-}
-
-export const getGeneralChartInteractions = (text: string[]) => {
-  const modeBarDescriptions = new Map([
-    ['Download plot as a png', `${text.includes('Download plot as a png') ? `${SvgIcons.CAMERA} <b>Screenshot</b>: You can download a .png of the chart.<br/>`: ''}`],
-    ['Zoom', `${text.includes('Zoom') ? `${SvgIcons.ZOOM} <b>Zooming</b>: Click and drag the mouse over a certain part of the visualization to zoom in and get a more detailed view of the data.<br/>`: ''}`],
-    ['Pan', `${text.includes('Pan') ? `${SvgIcons.PAN} <b>Panning</b>: You can move the view left and right while dragging the mouse.<br/>`: ''}`],
-    ['Box Select', `${text.includes('Box Select') ? `${SvgIcons.BOX_SELECTION} <b>Selection</b>: Drag the mouse over the chart to select some data.<br/>`: ''}`],
-    ['Lasso Select', `${text.includes('Lasso Select') ? `${SvgIcons.LASSO_SELECTION} <b>Lasso Select</b>: Select the desired data by drawing a lasso loop.<br/>`: ''}`],
-    ['Zoom in', `${text.includes('Zoom in') ? `${SvgIcons.ZOOM_IN} <b>Zoom in</b>: Zoom in to get a more detailed view of the chart.<br/>`: ''}`],
-    ['Zoom out', `${text.includes('Zoom out') ? `${SvgIcons.ZOOM_OUT} <b>Zoom out</b>: Zoom out to get a better overview of the chart.<br/>`: ''}`],
-    ['Autoscale', `${text.includes('Autoscale') ? `${SvgIcons.AUTO_SCALE} <b>Autoscale</b>: Zooms out to get a view of the whole visualization.<br/>`: ''}`],
-    ['Reset axes', `${text.includes('Reset axes') ? `${SvgIcons.RESET} <b>Reset</b>: It takes the chart to the inital layout settings.<br/>`: ''}`]
-  ]) 
-  return modeBarDescriptions
-}
-
-export const getModeBarMessages = (modebarInteractions: Map<string,string>) => { 
-  const cameraIcon = modebarInteractions.get('Download plot as a png') || '';  
-  const zoomIcon = modebarInteractions.get('Zoom') || '';
-  const panIcon = modebarInteractions.get('Pan') || '';
-  const selectionIcon = modebarInteractions.get('Box Select') || '';
-  const lassoSelectIcon = modebarInteractions.get('Lasso Select') || '';
-  const zoomInIcon = modebarInteractions.get('Zoom in') || '';
-  const zoomOutIcon = modebarInteractions.get('Zoom out') || '';
-  const autoScaleIcon = modebarInteractions.get('Autoscale') || '';
-  const resetIcon = modebarInteractions.get('Reset axes') || '';  
-  return {cameraIcon, zoomIcon, panIcon, selectionIcon, lassoSelectIcon, zoomInIcon, zoomOutIcon, autoScaleIcon, resetIcon};
 }
